requirements.txt
streamlit
groq
pytesseract
Pillow
opencv-python
pyzbar
requests
supabase
python-dotenv
numpy
